+++ src/modules/servicos_promocao/servicos_promocao.entity.ts
@@ -23,7 +23,7 @@ export class ServicosPromocao extends BaseEntity {
   @JoinColumn()
   servico: Servicos;
 
-  @Column({ nullable: true, type: 'integer', default: 0 })
+  @Column({ nullable: true, type: 'decimal', default: 0 })
   precoPromocional: number;
 
   @ManyToOne(() => Promocoes, {
